requirements.txt
ipython~=9.0.1
ipywidgets~=8.1.5
keras~=3.9.0
pillow~=11.1.0
tensorflow~=2.16.2
ipympl
seaborn
jupyterlab
scikit-learn
h5py
